tools/scraper/requirements.txt
python-jobspy
numpy>=2.4.1
